add initial player stats to the response body

+++ FantasyWeb.Common/Models/TeamStats.cs
@@ -3,8 +3,8 @@
     public class TeamStats
     {
         public int TeamID { get; set; }
-        public double TeamGoalsForm { get; set; }
-        public double TeamGoalsAwayForm { get; set; }
+        public decimal TeamGoalsForm { get; set; }
+        public decimal TeamGoalsAwayForm { get; set; }
         public string TeamForm { get; set; } = string.Empty;
     }
 }
